backend/functions/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   starlette
blinker==1.9.0
    # via flask
cachecontrol==0.14.4
    # via firebase-admin
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   functions-framework
    #   uvicorn
cloudevents==1.12.0
    # via
    #   firebase-functions
    #   functions-framework
cryptography==46.0.4
    # via
    #   google-auth
    #   pyjwt
deprecation==2.1.0
    # via cloudevents
distro==1.9.0
    # via google-genai
firebase-admin==7.1.0
    # via
    #   modpocket-functions (pyproject.toml)
    #   firebase-functions
firebase-functions==0.5.0
    # via modpocket-functions (pyproject.toml)
flask==3.1.2
    # via
    #   firebase-functions
    #   flask-cors
    #   functions-framework
flask-cors==6.0.2
    # via firebase-functions
functions-framework==3.10.0
    # via firebase-functions
google-api-core==2.29.0
    # via
    #   firebase-admin
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth==2.48.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
    #   google-genai
google-cloud-core==2.5.0
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.23.0
    # via
    #   firebase-admin
    #   firebase-functions
google-cloud-storage==3.9.0
    # via firebase-admin
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-events==0.5.0
    # via firebase-functions
google-genai==1.62.0
    # via modpocket-functions (pyproject.toml)
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.78.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.78.0
    # via google-api-core
gunicorn==25.0.2
    # via
    #   functions-framework
    #   uvicorn-worker
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   firebase-admin
    #   google-genai
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
msgpack==1.1.2
    # via cachecontrol
packaging==26.0
    # via
    #   deprecation
    #   gunicorn
pillow==12.1.0
    # via modpocket-functions (pyproject.toml)
proto-plus==1.27.1
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   google-events
protobuf==6.33.5
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   google-events
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via google-genai
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.11.0
    # via
    #   firebase-admin
    #   firebase-functions
pyyaml==6.0.3
    # via firebase-functions
requests==2.32.5
    # via
    #   modpocket-functions (pyproject.toml)
    #   cachecontrol
    #   google-api-core
    #   google-auth
    #   google-cloud-storage
    #   google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via google-genai
starlette==0.52.1
    # via functions-framework
tenacity==9.1.4
    # via google-genai
typing-extensions==4.15.0
    # via
    #   firebase-functions
    #   google-genai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
uvicorn==0.40.0
    # via
    #   functions-framework
    #   uvicorn-worker
uvicorn-worker==0.4.0
    # via functions-framework
watchdog==6.0.0
    # via functions-framework
websockets==15.0.1
    # via google-genai
werkzeug==3.1.5
    # via
    #   flask
    #   flask-cors
    #   functions-framework
